GATEWAY-2954: Stubbing out classes for Direct Core Service
